install_language.sh: support Flatpak installation of Steam
Installing via Flatpak is a popular method of installing Steam.

Add support for the Steam's directory, when it is installed via Flatpak
to the function get_save_game_path().

Notes
-----
Some documentation about the directory structure of Flatpak apps:

  - https://docs.flatpak.org/en/latest/sandbox-permissions.html#reserved-paths

The FLATPAK_ID for Steam is `com.valvesoftware.Steam`:

  - https://flathub.org/apps/com.valvesoftware.Steam
  - https://github.com/flathub/com.valvesoftware.Steam

The directory `~/.var/app/com.valvesoftware.Steam` is also mentioned in
the wiki of the GitHub repository for Steam:

  - https://github.com/flathub/com.valvesoftware.Steam/wiki#the-application-is-using-a-lot-of-disk-space
  - https://github.com/flathub/com.valvesoftware.Steam/wiki#notes