Added a seperate IJ version of the custom dictionary
Dutch has a weird Digraph Ĳ which even Scrabble has flip-flopped on in or excluding. Technically it's a single letter, unicode and Word Play support it as a single character, but a lot of people see it as two seperate letters. So I included a version with the digraph as standard with it as a seperate tile, but also included a version that has them as seperate letters.
There are also house roules of Scrabble that allow using an Y tile as a Ĳ when it doesn't have an Ĳ tile. However I chose not to include that since then you can just use the version with the digrapgh tile.